+++ cov.sh
@@ -13,11 +13,6 @@ fi
 
 # export RUSTFLAGS="-C link-dead-code"
 
-cargo build
-if [[ $? != 0 ]]; then
-  exit 1
-fi
-
 cargo build --tests
 if [[ $? != 0 ]]; then
   exit 1
@@ -27,7 +22,7 @@ files=(target/debug/blkar)
 
 #for file in target/debug/blkar-*[^\.d]; do
 # for file in ${files[@]}; do
-for file in target/debug/blkar-*; do if [[ $file == *.d ]]; then continue; fi
+for file in target/debug/blkar_lib-*; do if [[ $file == *.d ]]; then continue; fi
   # mkdir -p "target/cov/$(basename $file)"
   mkdir -p "target/cov/blkar"
   kcov --exclude-pattern=/.cargo,/usr/lib --verify "target/cov/blkar" "$file"