+++ src/cli_decode.rs
@@ -17,17 +17,23 @@ pub fn sub_command<'a, 'b>() -> App<'a, 'b> {
         .arg(in_file_arg()
              .help("SBX container to decode"))
         .arg(out_arg()
-             .help("Decoded file name. Supply - to use STDOUT as output. If OUT is not provided,
-then the original file name stored in the SBX container (STOREDNAME) is used
-if present. If OUT is provided and is a directory, then the output file is
-stored as OUT/STOREDNAME if STOREDNAME is present. If OUT is provided and is
-not a directory, then it is used directly."))
+             .help("Decoded file name. Supply - to use STDOUT as output. Use ./- for files named -.
+If OUT is not provided, then the original file name stored in the SBX container
+(STOREDNAME) is used if present. If OUT is provided and is a directory, then
+the output file is stored as OUT/STOREDNAME if STOREDNAME is present. If OUT is
+provided and is not a directory, then it is used directly."))
         .arg(Arg::with_name("force")
              .short("f")
              .long("force")
              .help("Force overwrite even if OUT exists"))
         .arg(no_meta_arg())
         .arg(pr_verbosity_level_arg())
+        .arg(burst_arg()
+             .help("Burst error resistance level used by the container.
+Use this if the level used by the container is above 1000,
+as blkar will only guess up to 1000. Or use this when blkar
+fails to guess correctly. blkar uses this value only if
+output is stdout and container version is RS enabled."))
         .arg(verbose_arg()
              .help("Show reference block info"))
         .arg(json_arg())
@@ -77,6 +83,8 @@ pub fn decode<'a>(matches : &ArgMatches<'a>) -> i32 {
 
     let pr_verbosity_level = get_pr_verbosity_level!(matches, json_printer);
 
+    let burst = get_burst_opt!(matches, json_printer);
+
     let in_file = get_in_file!(matches, json_printer);
 
     let param = Param::new(get_ref_block_choice!(matches),
@@ -85,7 +93,8 @@ pub fn decode<'a>(matches : &ArgMatches<'a>) -> i32 {
                            in_file,
                            out,
                            matches.is_present("verbose"),
-                           pr_verbosity_level);
+                           pr_verbosity_level,
+                           burst);
     match decode_core::decode_file(&param) {
         Ok(Some(s)) => exit_with_msg!(ok json_printer => "{}", s),
         Ok(None)    => exit_with_msg!(ok json_printer => ""),